Update on "Reland [quant][tests] Enable tests to run on all qengine backends"
Summary:
Refactor tests to use supported_qengines

Test Plan:
python test/test_quantization.py

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
