+++ test/quantization/test_quantized_op.py
@@ -18,7 +18,7 @@ hu.assert_deadline_disabled()
 
 from torch.testing._internal.common_utils import TestCase
 from torch.testing._internal.common_quantized import _quantize, _dequantize, _calculate_dynamic_qparams, \
-    override_quantized_engine, supported_qengines
+    override_quantized_engine, supported_qengines, override_qengines
 
 np_dtype = {
     torch.quint8 : np.uint8,
@@ -1719,119 +1719,116 @@ class TestDynamicQuantizedLinear(TestCase):
         use_bias=st.booleans(),
         use_relu=st.booleans(),
         use_multi_dim_input=st.booleans(),
-        use_channelwise=st.booleans(),
-        qengine=st.sampled_from(("qnnpack", "fbgemm")))
+        use_channelwise=st.booleans())
+    @override_qengines
     def test_qlinear(self, batch_size, input_channels, output_channels,
-                     use_bias, use_relu, use_multi_dim_input, use_channelwise, qengine):
+                     use_bias, use_relu, use_multi_dim_input, use_channelwise):
 
-        if qengine not in supported_qengines:
-            return
-        if qengine == 'qnnpack':
+        if torch.backends.quantized.engine == 'qnnpack':
             use_channelwise = False
             use_relu = False
 
-        with override_quantized_engine(qengine):
-            qlinear_prepack = torch.ops.quantized.linear_prepack
-            if use_relu:
-                qlinear_dynamic = torch.ops.quantized.linear_relu_dynamic
-            else:
-                qlinear_dynamic = torch.ops.quantized.linear_dynamic
+        qlinear_prepack = torch.ops.quantized.linear_prepack
+        if use_relu:
+            qlinear_dynamic = torch.ops.quantized.linear_relu_dynamic
+        else:
+            qlinear_dynamic = torch.ops.quantized.linear_dynamic
 
-            if use_multi_dim_input:
-                batch_size *= 3  # Test the multi-dim input tensor
+        if use_multi_dim_input:
+            batch_size *= 3  # Test the multi-dim input tensor
 
-            X_scale = 1.0
-            X_zp = 0
-            X_value_min = 0
-            X_value_max = 255
-            X_q0 = np.round(np.random.rand(batch_size, input_channels) *
-                            (X_value_max - X_value_min)
-                            + X_value_min
-                            ).astype(np.uint8)
-            X_q0 = np.round(np.random.rand(batch_size, input_channels) *
-                            (X_value_max - X_value_min) + X_value_min).astype(np.uint8)
-            X_q0[0, 0] = X_value_min
-            X_q0[0, 1] = X_value_max
+        X_scale = 1.0
+        X_zp = 0
+        X_value_min = 0
+        X_value_max = 255
+        X_q0 = np.round(np.random.rand(batch_size, input_channels) *
+                        (X_value_max - X_value_min)
+                        + X_value_min
+                        ).astype(np.uint8)
+        X_q0 = np.round(np.random.rand(batch_size, input_channels) *
+                        (X_value_max - X_value_min) + X_value_min).astype(np.uint8)
+        X_q0[0, 0] = X_value_min
+        X_q0[0, 1] = X_value_max
 
-            # W_scale = 1.0
-            # W_zp = 0
-            W_scales = np.ones(output_channels)
-            W_zps = np.zeros(output_channels).astype(np.int)
-            W_value_min = -128
-            W_value_max = 127
-            W_q0 = np.round(
-                np.random.rand(output_channels, input_channels)
-                * (W_value_max - W_value_min)
-                + W_value_min
-            ).astype(np.int8)
-            W_q0[0, 0] = W_value_min
-            W_q0[1, 0] = W_value_max
+        # W_scale = 1.0
+        # W_zp = 0
+        W_scales = np.ones(output_channels)
+        W_zps = np.zeros(output_channels).astype(np.int)
+        W_value_min = -128
+        W_value_max = 127
+        W_q0 = np.round(
+            np.random.rand(output_channels, input_channels)
+            * (W_value_max - W_value_min)
+            + W_value_min
+        ).astype(np.int8)
+        W_q0[0, 0] = W_value_min
+        W_q0[1, 0] = W_value_max
 
-            b_value_min = -10
-            b_value_max = 10
-            b_q0 = np.round(
-                np.random.rand(output_channels) *
-                (b_value_max - b_value_min) + b_value_min
-            ).astype(np.int32) if use_bias else None
+        b_value_min = -10
+        b_value_max = 10
+        b_q0 = np.round(
+            np.random.rand(output_channels) *
+            (b_value_max - b_value_min) + b_value_min
+        ).astype(np.int32) if use_bias else None
 
-            if qengine == 'fbgemm':
-                avoid_vpmaddubsw_overflow_linear(
-                    batch_size,
-                    input_channels,
-                    output_channels,
-                    X_q0,
-                    X_value_min,
-                    X_value_max,
-                    W_q0,
-                    W_value_min,
-                    W_value_max,
-                )
+        if torch.backends.quantized.engine == 'fbgemm':
+            avoid_vpmaddubsw_overflow_linear(
+                batch_size,
+                input_channels,
+                output_channels,
+                X_q0,
+                X_value_min,
+                X_value_max,
+                W_q0,
+                W_value_min,
+                W_value_max,
+            )
 
-            X_fp32 = torch.from_numpy(_dequantize(X_q0, X_scale, X_zp)).to(dtype=torch.float)
-            if use_multi_dim_input:
-                X_fp32 = X_fp32.view(3, int(batch_size / 3), input_channels)
+        X_fp32 = torch.from_numpy(_dequantize(X_q0, X_scale, X_zp)).to(dtype=torch.float)
+        if use_multi_dim_input:
+            X_fp32 = X_fp32.view(3, int(batch_size / 3), input_channels)
 
-            # W_scale, W_zp = _calculate_dynamic_qparams(W_fp32, torch.qint8)
-            # We currently only check the case where W_scale = 1.0, W_zp = 0.
+        # W_scale, W_zp = _calculate_dynamic_qparams(W_fp32, torch.qint8)
+        # We currently only check the case where W_scale = 1.0, W_zp = 0.
 
-            if use_channelwise:
-                W_fp32 = torch.from_numpy(_dequantize(W_q0, W_scales.reshape(
-                    (-1, 1)), W_zps.reshape((-1, 1)))).to(dtype=torch.float)
-                W_q = torch.quantize_per_channel(W_fp32, scales=torch.from_numpy(W_scales),
-                                                 zero_points=torch.from_numpy(W_zps), axis=0, dtype=torch.qint8)
-                b_fp32 = torch.from_numpy(
-                    _dequantize(b_q0, X_scale * W_scales, 0)
-                ).to(dtype=torch.float) if use_bias else None
-            else:
-                W_fp32 = torch.from_numpy(_dequantize(
-                    W_q0, W_scales[0], W_zps[0])).to(dtype=torch.float)
-                W_q = torch.quantize_per_tensor(W_fp32, scale=W_scales[0], zero_point=(
-                    W_zps[0].astype(int).item()), dtype=torch.qint8)
-                b_fp32 = torch.from_numpy(
-                    _dequantize(b_q0, X_scale * int(W_scales[0].item()), 0)
-                ).to(dtype=torch.float) if use_bias else None
+        if use_channelwise:
+            W_fp32 = torch.from_numpy(_dequantize(W_q0, W_scales.reshape(
+                (-1, 1)), W_zps.reshape((-1, 1)))).to(dtype=torch.float)
+            W_q = torch.quantize_per_channel(W_fp32, scales=torch.from_numpy(W_scales),
+                                             zero_points=torch.from_numpy(W_zps), axis=0, dtype=torch.qint8)
+            b_fp32 = torch.from_numpy(
+                _dequantize(b_q0, X_scale * W_scales, 0)
+            ).to(dtype=torch.float) if use_bias else None
+        else:
+            W_fp32 = torch.from_numpy(_dequantize(
+                W_q0, W_scales[0], W_zps[0])).to(dtype=torch.float)
+            W_q = torch.quantize_per_tensor(W_fp32, scale=W_scales[0], zero_point=(
+                W_zps[0].astype(int).item()), dtype=torch.qint8)
+            b_fp32 = torch.from_numpy(
+                _dequantize(b_q0, X_scale * int(W_scales[0].item()), 0)
+            ).to(dtype=torch.float) if use_bias else None
 
-            # Observe X_fp32 and determine X_scale and X_zero_point, this should match
-            # internals of dynamic linear.
-            X_scale, X_zp = _calculate_dynamic_qparams(X_fp32, torch.quint8)
-            X_q = torch.quantize_per_tensor(X_fp32, scale=X_scale, zero_point=X_zp, dtype=torch.quint8)
+        # Observe X_fp32 and determine X_scale and X_zero_point, this should match
+        # internals of dynamic linear.
+        X_scale, X_zp = _calculate_dynamic_qparams(X_fp32, torch.quint8)
+        X_q = torch.quantize_per_tensor(X_fp32, scale=X_scale, zero_point=X_zp, dtype=torch.quint8)
 
-            # Weight prepacking operator for dynamic quantized Linear
-            W_prepack = qlinear_prepack(W_q, b_fp32)
-            # Dynamic quantized Linear operator with prepacked weight
-            Y_fp32 = qlinear_dynamic(X_q.dequantize(), W_prepack)
-            # Y_fp32 = qlinear_dynamic(X_fp32, W_prepack, b_fp32)
+        # Weight prepacking operator for dynamic quantized Linear
+        W_prepack = qlinear_prepack(W_q, b_fp32)
+        # Dynamic quantized Linear operator with prepacked weight
+        Y_fp32 = qlinear_dynamic(X_q.dequantize(), W_prepack)
+        # Y_fp32 = qlinear_dynamic(X_fp32, W_prepack, b_fp32)
 
-            Y_fp32_ref = F.linear(X_q.dequantize(), W_q.dequantize(), b_fp32)
-            # Y_fp32_ref = F.linear(X_fp32, W_fp32, b_fp32)
-            # if use_multi_dim_input:
-            #     Y_fp32_ref = Y_fp32_ref.view(3, int(batch_size / 3), output_channels)
+        Y_fp32_ref = F.linear(X_q.dequantize(), W_q.dequantize(), b_fp32)
+        # Y_fp32_ref = F.linear(X_fp32, W_fp32, b_fp32)
+        # if use_multi_dim_input:
+        #     Y_fp32_ref = Y_fp32_ref.view(3, int(batch_size / 3), output_channels)
 
-            if use_relu:
-                Y_fp32_ref[Y_fp32_ref < 0.0] = 0.0
+        if use_relu:
+            Y_fp32_ref[Y_fp32_ref < 0.0] = 0.0
 
-            self.assertEqual(Y_fp32, Y_fp32_ref,
-                             message="torch.ops.quantized.linear_dynamic (fbgemm) results are off")
+        self.assertEqual(Y_fp32, Y_fp32_ref,
+                         message="torch.ops.quantized.linear_dynamic (fbgemm) results are off")
 
     @unittest.skipUnless('fbgemm' in torch.backends.quantized.supported_engines,
                          " Quantized operations require FBGEMM. FBGEMM is only optimized for CPUs"
@@ -1918,14 +1915,12 @@ class TestQuantizedLinear(unittest.TestCase):
            use_bias=st.booleans(),
            use_relu=st.booleans(),
            use_multi_dim_input=st.booleans(),
-           use_channelwise=st.booleans(),
-           qengine=st.sampled_from(("qnnpack", "fbgemm")))
+           use_channelwise=st.booleans())
+    @override_qengines
     def test_qlinear(self, batch_size, input_channels, output_channels, use_bias,
-                     use_relu, use_multi_dim_input, use_channelwise, qengine):
-        if qengine not in supported_qengines:
-            return
+                     use_relu, use_multi_dim_input, use_channelwise):
         decimal_val = 4
-        if qengine == 'qnnpack':
+        if torch.backends.quantized.engine == 'qnnpack':
             use_channelwise = False
             use_multi_dim_input = False
             # QNNPACK supports uint8 in the kernels. In the op we shift the int8
@@ -1934,155 +1929,151 @@ class TestQuantizedLinear(unittest.TestCase):
             # off by one results.
             decimal_val = 0
 
-        with override_quantized_engine(qengine):
-            qlinear_prepack = torch.ops.quantized.linear_prepack
+        qlinear_prepack = torch.ops.quantized.linear_prepack
+        if use_relu:
+            qlinear = torch.ops.quantized.linear_relu
+        else:
+            qlinear = torch.ops.quantized.linear
+        if use_multi_dim_input:
+            batch_size *= 3  # Test the multi-dim input tensor
+        X_scale = 1.5
+        X_zp = 5
+        X_value_min = 0
+        X_value_max = 225
+        X_q0 = np.round(
+            np.random.rand(batch_size, input_channels) *
+            (X_value_max - X_value_min)
+            + X_value_min
+        ).astype(np.uint8)
+        W_scales = np.random.rand(output_channels)
+        W_zps = np.round(np.random.rand(output_channels) * 100 - 50).astype(np.int)
+        W_value_min = -128
+        W_value_max = 127
+        W_q0 = np.round(
+            np.random.rand(output_channels, input_channels)
+            * (W_value_max - W_value_min)
+            + W_value_min
+        ).astype(np.int8)
+        b_value_min = -10
+        b_value_max = 10
+        b_q0 = np.round(
+            np.random.rand(output_channels) *
+            (b_value_max - b_value_min) + b_value_min
+        ).astype(np.int32) if use_bias else None
+        avoid_vpmaddubsw_overflow_linear(
+            batch_size,
+            input_channels,
+            output_channels,
+            X_q0,
+            X_value_min,
+            X_value_max,
+            W_q0,
+            W_value_min,
+            W_value_max,
+        )
+        X = torch.from_numpy(_dequantize(
+            X_q0, X_scale, X_zp)).to(dtype=torch.float)
+        X_q = torch.quantize_per_tensor(
+            X, scale=X_scale, zero_point=X_zp, dtype=torch.quint8)
+        if use_channelwise:
+            W = torch.from_numpy(_dequantize(W_q0, W_scales.reshape(
+                (-1, 1)), W_zps.reshape((-1, 1)))).to(dtype=torch.float)
+            W_q = torch.quantize_per_channel(W, scales=torch.from_numpy(W_scales),
+                                             zero_points=torch.from_numpy(W_zps), axis=0, dtype=torch.qint8)
+            b = torch.from_numpy(_dequantize(
+                b_q0, X_scale * W_scales, 0)).to(dtype=torch.float) if use_bias else None
+            b_q = torch.quantize_per_channel(b, scales=torch.from_numpy(X_scale * W_scales),
+                                             zero_points=torch.zeros(output_channels, dtype=torch.long),
+                                             axis=0, dtype=torch.qint32) if use_bias else None
+        else:
+            W = torch.from_numpy(_dequantize(
+                W_q0, W_scales[0], W_zps[0])).to(dtype=torch.float)
+            W_q = torch.quantize_per_tensor(W, scale=W_scales[0], zero_point=(
+                W_zps[0].astype(int).item()), dtype=torch.qint8)
+            b = torch.from_numpy(_dequantize(
+                b_q0, X_scale * (W_scales[0].item()), 0)).to(dtype=torch.float) if use_bias else None
+            b_q = torch.quantize_per_tensor(
+                b, scale=X_scale * (W_scales[0].item()), zero_point=0, dtype=torch.qint32) if use_bias else None
+        # Compare X_scale * W_scale * input_channels * X_value_max * W_value_max with
+        # Y_scale * 255 (max for uint8).
+        Y_scale = 125.1234
+        Y_zp = 5
+        # Weight prepacking operator for quantized Linear
+        float_bias = b if use_bias else None
+        W_prepack = qlinear_prepack(W_q, float_bias)
+        if use_multi_dim_input:
+            X_q = X_q.view(3, int(batch_size / 3), input_channels)
+        # Quantized Linear operator with prepacked weight
+        Y_q = qlinear(X_q, W_prepack, Y_scale, Y_zp)
+        if not use_channelwise:
+            # Test the per-tensor quantization only
+            # Reference quantized Linear operator
+            Y_q_ref = qlinear_ref(X_q0, X_scale, X_zp, W_q0,
+                                  W_scales[0], W_zps[0], b_q0, Y_scale, Y_zp)
             if use_relu:
-                qlinear = torch.ops.quantized.linear_relu
-            else:
-                qlinear = torch.ops.quantized.linear
+                Y_q_ref[Y_q_ref < Y_zp] = Y_zp
             if use_multi_dim_input:
-                batch_size *= 3  # Test the multi-dim input tensor
-            X_scale = 1.5
-            X_zp = 5
-            X_value_min = 0
-            X_value_max = 225
-            X_q0 = np.round(
-                np.random.rand(batch_size, input_channels) *
-                (X_value_max - X_value_min)
-                + X_value_min
-            ).astype(np.uint8)
-            W_scales = np.random.rand(output_channels)
-            W_zps = np.round(np.random.rand(output_channels) * 100 - 50).astype(np.int)
-            W_value_min = -128
-            W_value_max = 127
-            W_q0 = np.round(
-                np.random.rand(output_channels, input_channels)
-                * (W_value_max - W_value_min)
-                + W_value_min
-            ).astype(np.int8)
-            b_value_min = -10
-            b_value_max = 10
-            b_q0 = np.round(
-                np.random.rand(output_channels) *
-                (b_value_max - b_value_min) + b_value_min
-            ).astype(np.int32) if use_bias else None
-            avoid_vpmaddubsw_overflow_linear(
-                batch_size,
-                input_channels,
-                output_channels,
-                X_q0,
-                X_value_min,
-                X_value_max,
-                W_q0,
-                W_value_min,
-                W_value_max,
-            )
-            X = torch.from_numpy(_dequantize(
-                X_q0, X_scale, X_zp)).to(dtype=torch.float)
-            X_q = torch.quantize_per_tensor(
-                X, scale=X_scale, zero_point=X_zp, dtype=torch.quint8)
-            if use_channelwise:
-                W = torch.from_numpy(_dequantize(W_q0, W_scales.reshape(
-                    (-1, 1)), W_zps.reshape((-1, 1)))).to(dtype=torch.float)
-                W_q = torch.quantize_per_channel(W, scales=torch.from_numpy(W_scales),
-                                                 zero_points=torch.from_numpy(W_zps), axis=0, dtype=torch.qint8)
-                b = torch.from_numpy(_dequantize(
-                    b_q0, X_scale * W_scales, 0)).to(dtype=torch.float) if use_bias else None
-                b_q = torch.quantize_per_channel(b, scales=torch.from_numpy(X_scale * W_scales),
-                                                 zero_points=torch.zeros(output_channels, dtype=torch.long),
-                                                 axis=0, dtype=torch.qint32) if use_bias else None
-            else:
-                W = torch.from_numpy(_dequantize(
-                    W_q0, W_scales[0], W_zps[0])).to(dtype=torch.float)
-                W_q = torch.quantize_per_tensor(W, scale=W_scales[0], zero_point=(
-                    W_zps[0].astype(int).item()), dtype=torch.qint8)
-                b = torch.from_numpy(_dequantize(
-                    b_q0, X_scale * (W_scales[0].item()), 0)).to(dtype=torch.float) if use_bias else None
-                b_q = torch.quantize_per_tensor(
-                    b, scale=X_scale * (W_scales[0].item()), zero_point=0, dtype=torch.qint32) if use_bias else None
-            # Compare X_scale * W_scale * input_channels * X_value_max * W_value_max with
-            # Y_scale * 255 (max for uint8).
-            Y_scale = 125.1234
-            Y_zp = 5
-            # Weight prepacking operator for quantized Linear
-            float_bias = b if use_bias else None
-            W_prepack = qlinear_prepack(W_q, float_bias)
-            if use_multi_dim_input:
-                X_q = X_q.view(3, int(batch_size / 3), input_channels)
-            # Quantized Linear operator with prepacked weight
-            Y_q = qlinear(X_q, W_prepack, Y_scale, Y_zp)
-            if not use_channelwise:
-                # Test the per-tensor quantization only
-                # Reference quantized Linear operator
-                Y_q_ref = qlinear_ref(X_q0, X_scale, X_zp, W_q0,
-                                      W_scales[0], W_zps[0], b_q0, Y_scale, Y_zp)
-                if use_relu:
-                    Y_q_ref[Y_q_ref < Y_zp] = Y_zp
-                if use_multi_dim_input:
-                    Y_q_ref = np.reshape(
-                        Y_q_ref, (3, int(batch_size / 3), output_channels))
-                # Assert equal
-                np.testing.assert_array_almost_equal(Y_q_ref, Y_q.int_repr().numpy(), decimal=decimal_val)
-            # Test both per-tensor and per-channel quantization
-            # Reference quantized result from PyTorch Linear operator
-            W_fp32 = W_q.dequantize().to(dtype=torch.float)
-            X_fp32 = X_q.dequantize().to(dtype=torch.float)
-            b_fp32 = b_q.dequantize().to(dtype=torch.float) if use_bias else None
-            Y_fp32_ref = F.linear(X_fp32, W_fp32, b_fp32)
-            if use_relu:
-                Y_fp32_ref[Y_fp32_ref < 0.0] = 0.0
-            Y_q_ref2 = torch.quantize_per_tensor(
-                Y_fp32_ref, Y_scale, Y_zp, torch.quint8)
+                Y_q_ref = np.reshape(
+                    Y_q_ref, (3, int(batch_size / 3), output_channels))
             # Assert equal
-            np.testing.assert_array_almost_equal(
-                Y_q_ref2.int_repr().numpy(), Y_q.int_repr().numpy(), decimal=decimal_val)
+            np.testing.assert_array_almost_equal(Y_q_ref, Y_q.int_repr().numpy(), decimal=decimal_val)
+        # Test both per-tensor and per-channel quantization
+        # Reference quantized result from PyTorch Linear operator
+        W_fp32 = W_q.dequantize().to(dtype=torch.float)
+        X_fp32 = X_q.dequantize().to(dtype=torch.float)
+        b_fp32 = b_q.dequantize().to(dtype=torch.float) if use_bias else None
+        Y_fp32_ref = F.linear(X_fp32, W_fp32, b_fp32)
+        if use_relu:
+            Y_fp32_ref[Y_fp32_ref < 0.0] = 0.0
+        Y_q_ref2 = torch.quantize_per_tensor(
+            Y_fp32_ref, Y_scale, Y_zp, torch.quint8)
+        # Assert equal
+        np.testing.assert_array_almost_equal(
+            Y_q_ref2.int_repr().numpy(), Y_q.int_repr().numpy(), decimal=decimal_val)
 
     """Tests the correctness of the quantized::linear_unpack op."""
     @given(W=hu.tensor(shapes=hu.array_shapes(2, 2,),
                        qparams=hu.qparams(dtypes=torch.qint8)),
-           use_channelwise=st.booleans(),
-           qengine=st.sampled_from(("qnnpack", "fbgemm")))
-    def test_qlinear_unpack(self, W, use_channelwise, qengine):
-        if qengine not in supported_qengines:
-            return
+           use_channelwise=st.booleans())
+    @override_qengines
+    def test_qlinear_unpack(self, W, use_channelwise):
         if qengine == 'qnnpack':
             use_channelwise = False
 
-        with override_quantized_engine(qengine):
-            W, (W_scale, W_zp, torch_type) = W
-            if use_channelwise:
-                output_channels = W.shape[0]
-                W_scales = torch.rand(output_channels).to(torch.double)
-                W_zps = torch.round(torch.rand(output_channels)
-                                    * 100 - 50).to(torch.int64)
-            qlinear_prepack = torch.ops.quantized.linear_prepack
-            qlinear_unpack = torch.ops.quantized.linear_unpack
+        W, (W_scale, W_zp, torch_type) = W
+        if use_channelwise:
+            output_channels = W.shape[0]
+            W_scales = torch.rand(output_channels).to(torch.double)
+            W_zps = torch.round(torch.rand(output_channels)
+                                * 100 - 50).to(torch.int64)
+        qlinear_prepack = torch.ops.quantized.linear_prepack
+        qlinear_unpack = torch.ops.quantized.linear_unpack
 
-            W = torch.from_numpy(W)
-            if use_channelwise:
-                W_q = torch.quantize_per_channel(
-                    W, W_scales, W_zps, 0, dtype=torch_type)
-            else:
-                W_q = torch.quantize_per_tensor(W, scale=W_scale, zero_point=W_zp,
-                                                dtype=torch_type)
-            # Weight prepacking operator for quantized Linear
-            W_prepack = qlinear_prepack(W_q)
-            # Weight unpack operator for quantized Linear (Used for serialization)
-            W_q_origin = qlinear_unpack(W_prepack)[0]
-            # Assert equal
-            np.testing.assert_equal(W_q.int_repr(), W_q_origin.int_repr().numpy())
-            if use_channelwise:
-                np.testing.assert_array_almost_equal(np.float32(W_q.q_per_channel_scales().numpy()),
-                                                     np.float32(
-                                                         W_q_origin.q_per_channel_scales().numpy()),
-                                                     decimal=4)
-                np.testing.assert_equal(W_q.q_per_channel_zero_points(
-                ).numpy(), W_q_origin.q_per_channel_zero_points().numpy())
-            else:
-                np.testing.assert_equal(np.float32(
-                    W_q.q_scale()), np.float32(W_q_origin.q_scale()))
-                np.testing.assert_equal(
-                    W_q.q_zero_point(), W_q_origin.q_zero_point())
+        W = torch.from_numpy(W)
+        if use_channelwise:
+            W_q = torch.quantize_per_channel(
+                W, W_scales, W_zps, 0, dtype=torch_type)
+        else:
+            W_q = torch.quantize_per_tensor(W, scale=W_scale, zero_point=W_zp,
+                                            dtype=torch_type)
+        # Weight prepacking operator for quantized Linear
+        W_prepack = qlinear_prepack(W_q)
+        # Weight unpack operator for quantized Linear (Used for serialization)
+        W_q_origin = qlinear_unpack(W_prepack)[0]
+        # Assert equal
+        np.testing.assert_equal(W_q.int_repr(), W_q_origin.int_repr().numpy())
+        if use_channelwise:
+            np.testing.assert_array_almost_equal(np.float32(W_q.q_per_channel_scales().numpy()),
+                                                 np.float32(
+                                                     W_q_origin.q_per_channel_scales().numpy()),
+                                                 decimal=4)
+            np.testing.assert_equal(W_q.q_per_channel_zero_points(
+            ).numpy(), W_q_origin.q_per_channel_zero_points().numpy())
+        else:
+            np.testing.assert_equal(np.float32(
+                W_q.q_scale()), np.float32(W_q_origin.q_scale()))
+            np.testing.assert_equal(
+                W_q.q_zero_point(), W_q_origin.q_zero_point())
 
 class TestQuantizedConv(unittest.TestCase):
     def _test_qconv_unpack_impl(
@@ -2274,8 +2265,8 @@ class TestQuantizedConv(unittest.TestCase):
            Y_zero_point=st.integers(0, 4),
            use_bias=st.booleans(),
            use_relu=st.sampled_from([False]),
-           use_channelwise=st.booleans(),
-           qengine=st.sampled_from(("qnnpack", "fbgemm")))
+           use_channelwise=st.booleans())
+    @override_qengines
     def test_qconv2d(
             self,
             batch_size,
@@ -2300,11 +2291,8 @@ class TestQuantizedConv(unittest.TestCase):
             use_bias,
             use_relu,
             use_channelwise,
-            qengine
     ):
-        if qengine not in supported_qengines:
-            return
-        if qengine == 'qnnpack':
+        if torch.backends.quantized.engine == 'qnnpack':
             use_channelwise = False
 
         input_channels = input_channels_per_group * groups
@@ -2314,26 +2302,25 @@ class TestQuantizedConv(unittest.TestCase):
         pads = (pad_h, pad_w)
         dilations = (dilation, dilation)
 
-        with override_quantized_engine(qengine):
-            qconv = torch.ops.quantized.conv2d
-            if use_relu:
-                qconv = torch.ops.quantized.conv2d_relu
-            qconv_prepack = torch.ops.quantized.conv2d_prepack
-            conv_op = torch.nn.Conv2d(
-                input_channels,
-                output_channels,
-                kernels,
-                strides,
-                pads,
-                dilations,
-                groups,
-            )
-            self._test_qconv_impl(
-                qconv, qconv_prepack, conv_op, batch_size,
-                input_channels_per_group, (height, width),
-                output_channels_per_group, groups, kernels, strides, pads,
-                dilations, X_scale, X_zero_point, W_scale, W_zero_point,
-                Y_scale, Y_zero_point, use_bias, use_relu, use_channelwise)
+        qconv = torch.ops.quantized.conv2d
+        if use_relu:
+            qconv = torch.ops.quantized.conv2d_relu
+        qconv_prepack = torch.ops.quantized.conv2d_prepack
+        conv_op = torch.nn.Conv2d(
+            input_channels,
+            output_channels,
+            kernels,
+            strides,
+            pads,
+            dilations,
+            groups,
+        )
+        self._test_qconv_impl(
+            qconv, qconv_prepack, conv_op, batch_size,
+            input_channels_per_group, (height, width),
+            output_channels_per_group, groups, kernels, strides, pads,
+            dilations, X_scale, X_zero_point, W_scale, W_zero_point,
+            Y_scale, Y_zero_point, use_bias, use_relu, use_channelwise)
 
     """Tests the correctness of the quantized::qconv_unpack op."""
     @given(
@@ -2353,22 +2340,19 @@ class TestQuantizedConv(unittest.TestCase):
                                 zero_point_max=0)]),
         stride_h=st.integers(1, 3), stride_w=st.integers(1, 3),
         pad_h=st.integers(1, 2), pad_w=st.integers(1, 2),
-        channelwise=st.booleans(),
-        qengine=st.sampled_from(("qnnpack", "fbgemm")))
+        channelwise=st.booleans())
+    @override_qengines
     def test_qconv_unpack(
-        self, inputs, stride_h, stride_w, pad_h, pad_w, channelwise, qengine
+        self, inputs, stride_h, stride_w, pad_h, pad_w, channelwise
     ):
-        if qengine not in supported_qengines:
-            return
-        if qengine == 'qnnpack':
+        if torch.backends.quantized.engine == 'qnnpack':
             channelwise = False
 
-        with override_quantized_engine(qengine):
-            qconv_prepack = torch.ops.quantized.conv2d_prepack
-            qconv_unpack = torch.ops.quantized.conv2d_unpack
-            self._test_qconv_unpack_impl(
-                qconv_prepack, qconv_unpack, inputs, (stride_h, stride_w),
-                (pad_h, pad_w), channelwise)
+        qconv_prepack = torch.ops.quantized.conv2d_prepack
+        qconv_unpack = torch.ops.quantized.conv2d_unpack
+        self._test_qconv_unpack_impl(
+            qconv_prepack, qconv_unpack, inputs, (stride_h, stride_w),
+            (pad_h, pad_w), channelwise)
 
     """Tests the correctness of quantized 1D convolution op."""
     @given(batch_size=st.integers(1, 6),
@@ -2386,8 +2370,8 @@ class TestQuantizedConv(unittest.TestCase):
            W_zero_point=st.lists(st.integers(-5, 5), min_size=1, max_size=2),
            Y_scale=st.floats(4.2, 5.6),
            Y_zero_point=st.integers(0, 4),
-           use_bias=st.booleans(),
-           qengine=st.sampled_from(("qnnpack", "fbgemm")))
+           use_bias=st.booleans())
+    @override_qengines
     def test_qconv1d(
         self,
         batch_size,
@@ -2406,10 +2390,7 @@ class TestQuantizedConv(unittest.TestCase):
         Y_scale,
         Y_zero_point,
         use_bias,
-        qengine,
     ):
-        if qengine not in supported_qengines:
-            return
 
         input_channels = input_channels_per_group * groups
         output_channels = output_channels_per_group * groups
@@ -2436,27 +2417,26 @@ class TestQuantizedConv(unittest.TestCase):
             true_outp, scale=Y_scale, zero_point=Y_zero_point,
             dtype=torch.quint8)
 
-        with override_quantized_engine(qengine):
-            conv_op = torch.nn.quantized.Conv1d(
-                input_channels,
-                output_channels,
-                kernel,
-                stride,
-                pad,
-                dilation,
-                groups,
-            )
-            # Get the quantized weights and the output quantization params.
-            conv_op.set_weight_bias(W_q, bias_float)
-            conv_op.scale = float(Y_scale)
-            conv_op.zero_point = int(Y_zero_point)
+        conv_op = torch.nn.quantized.Conv1d(
+            input_channels,
+            output_channels,
+            kernel,
+            stride,
+            pad,
+            dilation,
+            groups,
+        )
+        # Get the quantized weights and the output quantization params.
+        conv_op.set_weight_bias(W_q, bias_float)
+        conv_op.scale = float(Y_scale)
+        conv_op.zero_point = int(Y_zero_point)
 
-            q_outp = conv_op(X_q)
+        q_outp = conv_op(X_q)
 
-            np.testing.assert_array_almost_equal(
-                q_result_ref.int_repr().numpy(),
-                q_outp.int_repr().numpy(),
-                decimal=0)
+        np.testing.assert_array_almost_equal(
+            q_result_ref.int_repr().numpy(),
+            q_outp.int_repr().numpy(),
+            decimal=0)
 
     @given(batch_size=st.integers(1, 4),
            input_channels_per_group=st.sampled_from([2, 4, 5, 8, 16]),
